refactor: made the method simpler.
The error message will be set on the step that failed, the ErrorMessage was the same message, so i removed it. The GetBackground was not needed because in the testresults, there is no background, the steps from the feature file will be duplicated on all scenarios. And at last the sorting function is not needed, they will match either way.

Refs: #10