feat: Handle more gherkin files
* refactor: removed some unused imports and cleaned up some.

* feat: for testing support for comments in .gurka files

Refs: #10

* feat: for testing markdown support.

Refs: #10

* feat: for testing a background support, testing both background in the root and also for background inside rules.

Refs: #10

* feat: example where tables are included.

Refs: #10

* feat: tables will now be added to the step in a format that should be able to render in markdown.

Refs: #10

* refactor: refactored and moved ParseTestOutput

I refactored the method to handle errors better. Before it looped 2 rows at a time, but the problem was that if errors occur, they are ofter longer than 2 lines.

Refs: #10

* refactor: if there are table, it will have their own property, this because it is easier to compare steps if the table is not included.

Refs: #10

* refactor: if there are a table, it will have their own property. Like the commit before this is needed for both the Scenario.cs and Background.cs.

Refs: #10

* chore: changed one test to fail, for testing how GenGurka handles it.

Refs: #10

* feat: made the Product.TestsPassed to be false or true depending on the features in it.

Refs: #10

* feat: added background to features and rules.

The Background.cs is very similar to a Scenario.cs, because backgrounds have steps themselves.

Refs: #10

* refactor: small refactorings, also Background.cs now has a GetSteps method.

Refs: #10

* feat: added GetSteps as a private method on Feature.cs.

This is because i need to check for steps in background and rules also, this could not be done in Scenario.cs like it was before.

Refs: #10

* refactor: removed ParseTestError and refactored the TestsPassed to take the results for the steps.

Refs: #10

* feat: remade the method that matches the feature with the scenario, this will now also check for background if there are no scenario matching.

Refs: #10

* refactor: made the method simpler.

The error message will be set on the step that failed, the ErrorMessage was the same message, so i removed it. The GetBackground was not needed because in the testresults, there is no background, the steps from the feature file will be duplicated on all scenarios. And at last the sorting function is not needed, they will match either way.

Refs: #10

* refactor: The testspassed was not right, but now it takes every step and checks.

Refs: #10

* refactor: small refactorings

Refs: #10