Test with production features
* feat: added header separation to table, to render correctly.

* feat: Added a resizeable sidebar where all the links to different features will be.

Refs: #7

* chore: added a font.

Read about this font here:
https://www.brailleinstitute.org/freefont/

This font is free to use and helps people with reading problems read text better with this font.

Refs: #7

* feat: added a basic layout that lists all the features in a project

Refs: #7

* feat: started with a page for a individual page.

Now i just got it working with navigating.

Refs: #7

* refactor: small changes

Refs: #7

* feat: moved the showing of features to the dynamic feature page.

Refs: #7

* chore: added MarkDig, markdown to html parser.

Refs: #7

* feature: styles for several features.

Refs: #7

* feat: added list of scenarios in sidemenu.

this way of listing all the scenarios was way better than having a dropdown on every feature-page.

Refs: #7

* feat: a start for searching features and keep sidebar width in localstorage.

Refs: #7

* refactor: renaming classes and now have the description at the start of the feature.

Refs: #7

* feat: more description support

Now supports descriptions inside backgrounds, scenarios, and these inside of rules.

Refs: #12

* feat: dropdown for a feature, do show all scenarios inside that one.

Refs: #7

* chore: small description on index page.

Refs: #7

* refactor: small changes and made when resizing sidebar you dont select text that is not intended.

Refs: #7

* feat: added a file-reader.

This file reader reads the latest .gurka file from the GurkaFiles directory, makes it very easy to add new runs to the project, just add it to the folder, restart and its done.

Refs: #7

* chore: removed config file, because now you can pass arguments.

Refs: #7

* feat: you can now pass the project name as argument.

this feature is good because it will have the project name in the .gurka file name and display it in VizGurka aswell.

Refs: #7

* chore: small refactorings.

Refs: #7

* feat: made the background appear a little more different than a scenario.

* chore: help for project name

Refs: #7

* feat: display time for scenario and feature.

Refs: #7

* feat: did a super simple demo of how the new proccess can work

* chore: upgraded packages for GenGurka and ReqnrollDemo.Spec

Refs: #14

* feat: now dotnet test runs auto inside console application.

Refs: #14

* feat: default directories and files.

Now default paths will be set if none is provided. This makes it easier to use.

Refs: #14

* docs: added some more info in README.md

Note, GenGurka needs to be published to nuget.org and setup so that you can install it as a tool.

Refs: #14

* feat: modifications for production run to work.

* feat: started rewriting to scenarios, steps and features to have a notImplemented status

* feat: continue status rewrite

* feat: sidebar scroll state in local storage.

This was something i was annoyed about, so now the scroll state of the sidebar will be saved when you change page.

* chore: added a icon for showing when features and scenarios are not implemented.

* refactor: refactored classes to use the new Status instead of TestsPassed

* feat/refactor: made the parse-method handle production output.

The production output came with some that was not need. Now the parseing is done correctly and also split the method into smaller bits.

* refactor: refactored the mapping between gherkindoc and features.

This refactor also includes support for the "@ignore" tag on features and scenarios. To set the status to NotImplemented.

* refactor: refactored this together with the mapping between gherkindoc and features.

* feat: added support for NotImplemented in VizGurka.

* feat: helper for setting the right icon based on status.

* feat: possibility for GenGurka to run tests, but these are commented out for now.